pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ccl"
version = "2.0.0"
description = "Compliance Checker Legalized — offline data compliance validation engine"
requires-python = ">=3.12"
dependencies = [
    "pandas>=2.2",
    "pyarrow>=15",
    "python-frontmatter>=1.1",
    "pyyaml>=6",
    "click>=8.1",
    "pydantic>=2.6",
]

[project.scripts]
ccl = "ccl.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--tb=short"
